AAFG-SBIR/requirements.txt
torch>=2.1
torchvision>=0.16
numpy>=1.24
pandas>=2.0
opencv-python>=4.8
Pillow>=10.0
tqdm>=4.66
tensorboard>=2.14
pyyaml>=6.0
scikit-learn>=1.3
pytest>=7.4
